Factor out data layer.
