+++ backend/chainlit/data/base.py
@@ -1,0 +1,109 @@
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Protocol, Union
+
+from chainlit.data.utils import queue_until_user_message
+from chainlit.types import (
+    Feedback,
+    PageInfo,
+    PaginatedResponse,
+    Pagination,
+    ThreadDict,
+    ThreadFilter,
+)
+from chainlit.user import PersistedUser, User
+
+if TYPE_CHECKING:
+    from chainlit.element import Element, ElementDict
+    from chainlit.step import StepDict
+
+
+class BaseDataLayer:
+    """Base class for data persistence."""
+
+    async def get_user(self, identifier: str) -> Optional["PersistedUser"]:
+        return None
+
+    async def create_user(self, user: "User") -> Optional["PersistedUser"]:
+        pass
+
+    async def delete_feedback(
+        self,
+        feedback_id: str,
+    ) -> bool:
+        return True
+
+    async def upsert_feedback(
+        self,
+        feedback: Feedback,
+    ) -> str:
+        return ""
+
+    @queue_until_user_message()
+    async def create_element(self, element: "Element"):
+        pass
+
+    async def get_element(
+        self, thread_id: str, element_id: str
+    ) -> Optional["ElementDict"]:
+        pass
+
+    @queue_until_user_message()
+    async def delete_element(self, element_id: str, thread_id: Optional[str] = None):
+        pass
+
+    @queue_until_user_message()
+    async def create_step(self, step_dict: "StepDict"):
+        pass
+
+    @queue_until_user_message()
+    async def update_step(self, step_dict: "StepDict"):
+        pass
+
+    @queue_until_user_message()
+    async def delete_step(self, step_id: str):
+        pass
+
+    async def get_thread_author(self, thread_id: str) -> str:
+        return ""
+
+    async def delete_thread(self, thread_id: str):
+        pass
+
+    async def list_threads(
+        self, pagination: "Pagination", filters: "ThreadFilter"
+    ) -> "PaginatedResponse[ThreadDict]":
+        return PaginatedResponse(
+            data=[],
+            pageInfo=PageInfo(hasNextPage=False, startCursor=None, endCursor=None),
+        )
+
+    async def get_thread(self, thread_id: str) -> "Optional[ThreadDict]":
+        return None
+
+    async def update_thread(
+        self,
+        thread_id: str,
+        name: Optional[str] = None,
+        user_id: Optional[str] = None,
+        metadata: Optional[Dict] = None,
+        tags: Optional[List[str]] = None,
+    ):
+        pass
+
+    async def delete_user_session(self, id: str) -> bool:
+        return True
+
+    async def build_debug_url(self) -> str:
+        return ""
+
+
+class BaseStorageClient(Protocol):
+    """Base class for non-text data persistence like Azure Data Lake, S3, Google Storage, etc."""
+
+    async def upload_file(
+        self,
+        object_key: str,
+        data: Union[bytes, str],
+        mime: str = "application/octet-stream",
+        overwrite: bool = True,
+    ) -> Dict[str, Any]:
+        pass
